+++ packages/locale-de/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cosmicdrift/kumiko-locale-de
 
+## 0.243.3
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-framework@0.243.3
+
 ## 0.243.2
 
 ### Patch Changes