+++ packages/renderer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cosmicdrift/kumiko-renderer
 
+## 0.243.3
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-framework@0.243.3
+- @cosmicdrift/kumiko-headless@0.243.3
+
 ## 0.243.2
 
 ### Patch Changes